Fixed bug with default values of enums
Some objects in the EP schema had no default value and not the option ""(empty string) in the enum list. For these rare cases, the first enum value is picked. It is left to the user to see that the default does not match. This is not a problem as this only affects important inputs where EP didn't think that a user would leave it to default.